This is a big release, building up to the major 1.0 release. Expect some issues, there will be 0.9.9.postX releases with just bug fixes between now and 1.0
Added the ability to pass in an extra ontology document which gets parsed and mixed with the
data graph before pre-inferencing. This helps in the cases where the target data graph contains a data snippet which
can only be fully expanded with the help of an external ontology document containing RDFS and OWL axioms.
- Use `ont_graph=path_to_graph` in the python module or
- Use `-e` or `--ont-graph` on the command line utility to take advantage of this feature.
SHACL graph or ONT graph can now be a Web URL, rather than a file path.
- This works from the module validator entrypoint or the commandline tool.
Added built in tests for issue#14 and for the commandline tool.
Added new details to the README about the above new features.
Added coverage statistics to the README.
Started adding some hopefully-informative debugging output messages when debug mode is turned on. More to come.
Pre-inferencing can now only ever be run once per Validator instance, this is an attempt to prevent running
pre-inferencing multiple times unnecessarily.
Internal shapes lookup cache is now stored in the `ShapesGraph` instance, rather than in a global static class
variable on the `Shape` class
Fixed some bugs in the examples code, thanks @johannesloetzsch!
Lots of code coverage specific changes, and comments where we can improve coverage.