Add ConstraintLoadError link
This patch adds a link that type review noted was missing.

The link added in this patch is duplicated from prior references in the
same file.  The SHACL JavaScript Extensions page seems to offer another
possible link, but I was not able to find a note there on the described
requirements.

References:
* https://www.w3.org/TR/shacl-js/#js-constraints

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>