Enable GitHub Actions
This patch is an attempt to enable GitHub Actions.  It is a reduced
implementation of the patch in PR 250, and intentionally doesn't run
tests.

References:
* https://github.com/RDFLib/pySHACL/pull/250

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>